requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
resend>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.6.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
ultralytics>=8.0.0
opencv-python>=4.8.0
easyocr>=1.7.0